// src/schemas/alimentar.schema.ts
import { z } from 'zod';

export const createPlanoAlimentarSchema = z.object({
  titulo: z.string().min(3, 'Título deve ter no mínimo 3 caracteres'),
  descricao: z.string().optional(),
  alunoId: z.string().uuid('ID do aluno inválido')
});

export type CreatePlanoAlimentarInput = z.infer<typeof createPlanoAlimentarSchema>;